# docs/v2/gateway-api.openapi.yaml
openapi: 3.0.3
info:
  title: Erebrus Gateway API v2
  version: 2.0.0
  description: |
    Public API of the Erebrus gateway (gateway.erebrus.io): auth, accounts,
    node discovery, VPN client provisioning, entitlements, organizations, and
    admin. Webapp: erebrus.io / dev.erebrus.io. FROZEN (v2.0): additive only.

    Auth modes:
      - User PASETO (wallet-signature or social login) — Authorization: Bearer
      - Org API key — X-Api-Key header (org-scoped routes; metered)
      - Node PASETO — node registration + WS upgrade only
      - Admin = user PASETO with role=admin claim
servers:
  - url: https://gateway.erebrus.io
paths:
...
  /api/v2/rank/claim:
    post:
      summary: Spend claimable XP on a reward
      description: "v2.0 reward: free_days (XP_FREE_DAYS_COST → XP_FREE_DAYS_GRANT days, source='rank', stacked on the current period). Logged with IP + device."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                type: object,
                required: [reward],
                properties: { reward: { type: string, enum: [free_days] } },
              }
      responses:
        "200": { description: "{ reward, xp_spent, days, subscription }" }
        "400": { description: Unknown reward }
        "409": { description: Not enough claimable XP }
...
  /api/v2/social/telegram:
    post:
      summary: Verify + link a Telegram account (Login Widget payload)
      description: "Body = the widget's key=value fields incl. hash. HMAC-verified against TELEGRAM_BOT_TOKEN. First link awards social_verified XP."
      responses:
        "200": { description: "{ provider, handle, newly_linked }" }
        "401": { description: Verification failed }
        "409": { description: Linked to another account }
        "503": { description: TELEGRAM_BOT_TOKEN not configured }
  /api/v2/social/x:
    post:
      summary: Verify + link an X account (OAuth2 access token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                type: object,
                required: [access_token],
                properties: { access_token: { type: string } },
              }
      responses:
        "200": { description: "{ provider, handle, newly_linked }" }
        "401": { description: Verification failed }
        "409": { description: Linked to another account }
...
  /api/v2/vpn/clients:
    get:
      summary: List my VPN clients
      responses: { "200": { description: Clients with node/region/status } }
    post:
      summary: Provision a VPN client on a node
      description: |
        Requires entitlement (active subscription/trial, or org API key).
        Flow: pending row committed → PUT node /api/v2/peers/{uuid} (10s
        timeout, 2 retries, idempotent) → active. Returns the full credential
        bundle: WireGuard conf (with REPLACE_WITH_PRIVATE_KEY placeholder),
        vless://, hysteria2://, and a complete sing-box profile with
        "auto" (urltest: wg→vless→hy2) and "stealth" (vless→hy2) groups.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, node_id, wg_public_key]
              properties:
                name: { type: string }
                node_id: { type: string }
                wg_public_key: { type: string }
                wg_preshared_key: { type: string }
                idempotency_key: { type: string }
      responses:
        "201": { description: Credential bundle }
        "402": { description: No active entitlement }
        "409": { description: Node draining or client limit reached }
        "502": { description: Node unreachable — no client created }
...
  /api/v2/subscriptions/trial:
    post:
      {
        summary: "Start the one-time free trial (7 days, plan 'pro')",
        responses:
          {
            "201": { description: Trial active },
            "409": { description: Already used },
          },
      }
  /api/v2/subscriptions/nft/refresh:
    post:
      summary: Verify the gating NFT and grant/refresh a 30-day entitlement
      description: >
        Grants 30 days directly regardless of trial state (a mid-trial user is
        upgraded, since the 30d NFT entitlement outlasts the 7d trial).
        Refreshable while held; one per user.
      responses:
        "200": { description: NFT entitlement granted/refreshed }
        "400": { description: No wallet on account }
        "403": { description: Wallet does not hold the required NFT }
        "502": { description: NFT ownership check failed }
        "503": { description: NFT gating not configured }

  # ───────────────────────── organizations ─────────────────────────
...
  /api/v2/admin/nodes/{id}/min_tier:
    {
      post:
        {
          summary: "Set a node's premium-pool tier gate",
          responses:
            {
              "200": { description: "{ node_id, min_tier }" },
              "404": { description: Node not found },
            },
        },
    }
...
  /api/v2/admin/perks/{id}/grant:
    {
      post:
        {
          summary: "Grant a perk to a user (by user_id or wallet)",
          responses:
            {
              "200": { description: Granted },
              "404": { description: No user for wallet },
            },
        },
    }
...
  /api/v2/admin/settings:
    get:
      summary: List platform_settings (XP, trials, rate limits, auth text, …)
      responses:
        "200": { description: "{ settings[{key, value, description}] }" }
    patch:
      summary: Update platform settings (hot-reloads rate limits, XP weights, PASETO TTL for new tokens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: object
                  additionalProperties: { type: string }
                  example: { trial_period: "168h", xp_referrer_points: "100" }
      responses:
        "200": { description: "{ status: updated }" }
        "400": { description: Unknown key or invalid value }
...